// src/types/profile.d.ts
export interface Profile {
  id: string;
  full_name: string;
  short_name: string;
  bio: string;
  
  email: string;
  whatsapp_number: string;
  url_linkedin: string;
  url_github: string;
  url_instagram: string;
  url_threads: string;
  cv_url: string;
  profile_image: string;
  avatar_url: string;
}